+++ MeshEditor/OneSimpleHandlerBase.cpp
@@ -25,7 +25,7 @@ void OneSimpleHandlerBase::get_fhs_on_boundary ()
 void OneSimpleHandlerBase::get_fhs_on_interface ()
 {
 	interface_fhs.clear ();
-	auto V_ENTITY_PTR = mesh->request_vertex_property<unsigned long> ("entityptr");
+	auto V_ENTITY_PTR = mesh->request_vertex_property<unsigned int> ("entityptr");
 
 	for (auto bf_it = mesh->bf_iter (); bf_it; ++bf_it){
 		auto vhs_on_fh = JC::get_adj_vertices_around_face (mesh, *bf_it);
